+++ route.h
@@ -11,7 +11,7 @@ class route {
 public:
     route();
     route(string set_airline, string set_airline_ID, string set_departure_id, string set_destionation_ID);
-    double get_distance();
+    long double get_distance();
     vector<route> parse_routes_from_file();
     route search_route(string, string, vector<route>);
     void print_route();
@@ -23,8 +23,7 @@ public:
     Airport destination;
     string destination_id;
     bool is_air;
-    double distance;
-    //double weight;
+    long double distance;
 
     
 private:
@@ -32,9 +31,6 @@ private:
     void set_airports_distance();
 
     //return the distance of the route
-
-
-    
 
     //helper function to split string
     void split(const string & s, string c, vector<string> & v);
